requirements.txt
google-generativeai==0.7.2
openai==1.56.0
python-dotenv
langchain
langchain-openai
langchain-google-genai
transformers==4.44.2
accelerate==0.34.2
groq==0.11.0
langchain-huggingface
langchain-groq
 httpx==0.27.2 
#  --force-reinstall --quiet
